memdisk: Change patch area to include the entire MDI
Previously, the patch area began after the beginning of
the MDI, but continued past its end.  This is difficult
to express with C structures, so we've simply changed
the patch area to include the entire MDI.  So it now
includes the MDI size and MEMDISK version info, which
we simply make constant to avoid modifying.

Signed-off-by: Shao Miller <shao.miller@yrdsb.edu.on.ca>