memdisk: Force 32-bit mBFT "safe hook" pointer field
When mstructs.h is used by MDISKCHK.COM, pointers have
a 16-bit size and thus the mBFT structure sizeof() is
smaller under OpenWatcom.  We now union this pointer
member with a uint32_t to force the field to at least
32 bits.

There is another pointer in this header file which is
already unioned with a uint32_t, due to that field's
dual-purpose.  That field is fine under OpenWatcom.

Signed-off-by: Shao Miller <shao.miller@yrdsb.edu.on.ca>