diskstart: if no partition info is available, use bsHidden
If we have no partition information available, use the bsHidden field
(which is set by the extlinux installer in the case of non-FAT).  This
gives at least a hope of working correctly (for < 2 TiB disks, at
least) with the stock Vista/Win7 MBR.

Also, add a check for partition type != 0.  This helps catch the case
when DS:SI points into all-zero memory.

Signed-off-by: H. Peter Anvin <hpa@zytor.com>