chain.c32: only pass partition number in DH, when grldr= is used
Add grldr= as boot parameter, so the partition number is only passed to
grub4dos grldr, when grldr= is used.

Currently the partition number is passed in DH unconditionally.
As consequence, grldr will set its root partition on which is searches its
menu.lst to (hd0), because -1 (whole drive) is passed in DH as root partition:
  chain.c32 file=/grldr

When the conditional check, this problem is solved:
  chain.c32 file=/grldr
Grub4dos will search all partitions for its menu.lst

When you want to set the root partition, use something like the following:
  chain.c32 hd0,2 grldr=/grldr
  chain.c32 fs grldr=/grldr

Signed-off-by: Gert Hulselmans <gerth@zytor.com>