libcom32: Introduce disk header and module
An effort to move significant portions out of chain.c32 and
into library is underway.  We begin by simply moving
SECTOR into a header and building a disk module.

Note that some of this work will essentially duplicate
some of what's found in Pierre-Alexandre Meyer's fine
work in com32/gpllib/disk/.  A difference would be the
licensing, of course.

Signed-off-by: Shao Miller <shao.miller@yrdsb.edu.on.ca>