libinstaller: Add --mbr, --active, and --force options
In order to consolidate the installer option-parsing system,
we add the --mbr (-m), --active (-a), and --force (-f)
options from the DOS, Win32, and mtools installers to
libinstaller.

Since the -m and -a options are not currently valid for the
Linux 'syslinux' and 'extlinux' installers, we add error
messages to both of these.  The -f option is allowed but
does nothing for them.

Signed-off-by: Shao Miller <shao.miller@yrdsb.edu.on.ca>