diskio: fix name for secpercyl field, remove unused type field
Fix the name for the incorrectly named "t" (track) field; it is in
fact the sectors/cylinder (secpercyl) field.  Furthermore, remove the
completely unused "type" field... right now the pointer to the
rdwr_sectors field acts as a proxy, and if we need more complex stuff
in the future it is probably better handled by pointing to an ops
structure.

Signed-off-by: H. Peter Anvin <hpa@zytor.com>