core: store both UI and DEFAULT; change DefaultLevel if UI fails
This is not yet complete by itself.  If both a UI and a DEFAULT are
specified, store the last use of each entry (allow overrides).
DefaultLevel indicates the preferred option for use by auto_boot.  If
UI fails to load with a file not found, change DefaultLevel to
use the DEFAULT entry.

This does not bounce from UI to DEFAULT at this time.