ACPI: Improving tables detection
In some conditions, XSDT pointer could be null. In such conditions, we
have to parse tables pointed by RSDP.

This patch remove the table parsing from xsdt to acpi.c.
If the XSDT detection fails, let's switch back to the RSDP entries.

This was the case under KVM whereas VirtualBox features an XSDT.