elflink: Don't reload the current EXEC_MODULE module
As the ldlinux.c32 ELF module is an executable (EXEC_MODULE) and not a
library (LIB_MODULE), we need to make sure we don't reload it when
another executable wants to resolve a symbol exported by ldlinux.

Whenever an EXEC_MODULE module is loaded its MODULE_MAIN function is
executed. In the case of ldlinux, the MODULE_MAIN function drops the
user at a command line. So, what currently happens is that if we load
an executable module, say menu.c32, which needs symbols exported by
ldlinux and ldlinux is already loaded and running, we'll reload
ldlinux when resolving undefined symbols from menu.c32 and end up
dropping the user at a command prompt and _not_ resolve the symbols
and return control to menu.c32's MODULE_MAIN function. ldlinux.c32
effectively steals control of the proces.

Signed-off-by: Matt Fleming <matt.fleming@linux.intel.com>