elflink, cli: TAB key now displays labels
Now, if the user hits the TAB key at a prompt a list of all labels to
complete the current label typed on the command-line is printed. If
nothing is typed at the command-line all labels are printed.

This feature is available in the asm command-line interface so lets
make it available in the C version.

Signed-off-by: Matt Fleming <matt.fleming@linux.intel.com>