elflink: Detect circular dependencies during module load
It's possible for the 'modules.dep' file to contain circular
dependencies where module A has a dependency on module B, and module B
has a depedency on module A.

Currently what happens in the face of circular dependencies when
trying to parse and load the modules in 'modules.dep' is that we get
stuck in a loop in module_load_dependencies(), presumably eventually
blowing the stack.

To fix this we maintain a singly-linked list of module names that are
in the process of having their dependencies loaded, but that have not
completed the loading. Everytime we begin loading the dependencies for
a new module we check to see if that module name is already on the
linked list. If it is, we've discovered a circular reference so we
refuse to load the dependencies and print a helpful message to the
user which includes the circular dependency chain.

It would be possible to cope with these circular references if we were
to skip loading the dependencies for any modules that are on the
linked list and defer resolving any module relocs until after all
module images have been loaded into memory. However, it's more than
likely that a circular reference is indicative of a bug and that the
correct fix would be to extract the symbols causing the circular
dependency into their own module. So we treat circular dependencies as
an error case for now.

Signed-off-by: Matt Fleming <matt.fleming@linux.intel.com>