ldlinux, cli: Add support for F-keys
The old asm command-line interface had support for printing files
containing help messages when the F-keys (F1-F10) were pressed. Add
this support to ldlinux.

The config parser already had support for parsing config files with
F-key directives and the cli code already had support for executing
callbacks when one of the F-keys was hit, so this patch simply glues
all the pieces together and provides a function to cat the help file.

Signed-off-by: Matt Fleming <matt.fleming@linux.intel.com>