elflink: Jump to kaboom if we fail to load ldlinux.c32
Previously if we failed to load ldlinux.c32, we would drop out of
load_env32() and try to execute the assembly kernel loader.

With this patch, if we drop out of load_env32() we jump to kaboom
which drops the user at a command-line, at least that way they have a
chance of fixing the reason ldlinux.c32 failed to load.

Also, delete the comment that talks about loading and parsing the
configuration file - that is now done by ldlinux.c32.

Signed-off-by: Matt Fleming <matt.fleming@linux.intel.com>