Makefile: Move Makefile fragments into mk/
Move the MCONFIG files into a mk/ directory and give them more
descriptive names.

This is purely a cosmetic change to make the 'include' directives a
bit more coherent by making it obvious exactly which MCONFIG file
we're including. For example, in com32/lua/src/Makefile we exchange
the line,

	include ../../MCONFIG

for the much more comprehensible,

	include $(MAKEDIR)/com32.mk

Signed-off-by: Matt Fleming <matt.fleming@linux.intel.com>