ldlinux: Pass entire cmdline to execute()
We need to pass the entire cmdline to execute(), not just the kernel
name. execute() does all the required processing of cmdline arguments
and passes them to the loader in the correct format, e.g. for COM32
modules the cmdline arguments need to be passed in the traditional
char *argv[] way.

So stop using strtok() which modifies its first argument and means we
therefore lose the cmdline after the kernel name.

Previously, the "argc" and "argv" arguments that are passed to a
module's main() function contained bogus values.

Signed-off-by: Matt Fleming <matt.fleming@linux.intel.com>