thread: thread-switch the real-mode stack, too
When this code was originally written, we didn't have lmalloc().  Now
when lmalloc() is implemented, let each real-mode task have its own
stack.

Note that this means we absolutely have to continue to support the
SS != CS, DS model in the real-mode code, which should already be the
case, but...

Signed-off-by: H. Peter Anvin <hpa@zytor.com>