pxe: use the unified URL parsing framework
Use the unified URL parsing framework for TFTP and HTTP.  This should
also make it easier to add new protocols (e.g. FTP) in the near
future.

Note that HTTP redirects are still handled wrong: they really should
be sent all the way back to the top of URL parsing; there are sites in
the field which redirect to FTP URLs, for example.

Signed-off-by: H. Peter Anvin <hpa@zytor.com>