ldlinux: Fix "prompt" config option parsing
Previously, we were looking for the value of the prompt option 8
characters after the first 'p', clearly since "prompt" is only 6
characters we should be searching after the sixth character.

I noticed this bug when specifying "prompt 1" in my config file and
Syslinux still tried to boot the default kernel.

Signed-off-by: Matt Fleming <matt.fleming@linux.intel.com>