efi: Add UEFI firmware backend
This commit adds 32-bit support for a Unified Extensible Firmware
Interface backend.

We use a wrapper program to create an EFI exe with only the minimum
number of sections becaues the EFI loader doesn't like certain section
sections, such as .gnu.hash, which we require for resolving symbols
when ELF modules are loaded.

We're currently including a lot more symbols than are necessary in
efi/main.c so that the code compiles. Some of them are BIOS specific
but I haven't got around to fixing up the core to not reference them
directly yet.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>