core: Delete the aux segment
We don't need to use a special aux segment because we can represent
'fontbuf' with OFFS() and SEG(). We're guaranteed to be able to break
the 32-bit pointer up with these macros because we control where in
the address space the core is loaded. (Not all 32-bit pointers can be
represented with OFFS() and SEG(), only addresses in the range
0..0xFFFFF.)

This fixes the breakage that was introduced in commit 14531c47bc95
("core: Delete code that is duplicated in ldlinux").

This allows the default font to be displayed. Previously junk was
being returned in the COMBOOT API call to query the userfont, leading
the caller to believe that a user font was installed even when it
wasn't.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>